fixed compile warnings on Linux

+++ glib-core/unicode.cpp
@@ -114,7 +114,7 @@ void TUniCodec::TestUtf8(bool decode, size_t expectedRetVal, bool expectedThrow,
 		if (f) {
 			fprintf(f, "\n -> dest:    "); for (int i = 0; i < dest.Len(); i++) fprintf(f, (decode ? " %x" :  " %02x"), uint(dest[i]));
 			fprintf(f, "\n    expDest  "); for (int i = 0; i < expectedDest.Len(); i++) fprintf(f, (decode ? " %x" :  " %02x"), uint(expectedDest[i]));
-			fprintf(f, "\n    retVal = %d (expected %d)\n", retVal, expectedRetVal); }
+			fprintf(f, "\n    retVal = %d (expected %d)\n", (int) retVal, (int) expectedRetVal); }
 		if (retVal != expectedRetVal)
 			printf("!!!");
 		IAssert(retVal == expectedRetVal); IAssert(! expectedThrow);
@@ -317,7 +317,7 @@ void TUniCodec::TestUtf16(bool decode, size_t expectedRetVal, bool expectedThrow
 			if (f) {
 				fprintf(f, "\n -> dest:    "); for (int i = 0; i < dest.Len(); i++) fprintf(f, (decode ? " %x" :  fmt), uint(dest[i]));
 				fprintf(f, "\n    expDest  "); for (int i = 0; i < ed.Len(); i++) fprintf(f, (decode ? " %x" :  fmt), uint(ed[i]));
-				fprintf(f, "\n    retVal = %d (expected %d)\n", retVal, expectedRetVal); }
+				fprintf(f, "\n    retVal = %d (expected %d)\n", (int) retVal, (int) expectedRetVal); }
 			bool ok = true;
 			if (retVal != expectedRetVal) ok = false;
 			if (dest.Len() != ed.Len()) ok = false;
